+++ src/main/java/com/github/command17/fabricdiscordbridge/config/BotConfig.java
@@ -1,17 +1,51 @@
 package com.github.command17.fabricdiscordbridge.config;
 
-import com.github.command17.fabricdiscordbridge.config.entry.BoolConfigEntry;
-import com.github.command17.fabricdiscordbridge.config.entry.StringConfigEntry;
-import net.fabricmc.loader.api.FabricLoader;
+import de.maxhenkel.configbuilder.ConfigBuilder;
+import de.maxhenkel.configbuilder.entry.ConfigEntry;
+import de.maxhenkel.configbuilder.entry.EnumConfigEntry;
+import net.dv8tion.jda.api.entities.Activity;
+import org.jspecify.annotations.NonNull;
+import org.jspecify.annotations.Nullable;
+
+import java.util.function.Function;
 
 public class BotConfig {
-    private final ConfigBuilder builder = new ConfigBuilder("bot")
-            .setParentPath(FabricLoader.getInstance().getGameDir());
+    public final ConfigEntry<String> token;
+    public final ConfigEntry<Boolean> autoStart;
+    public final EnumConfigEntry<ActivityType> activityType;
+    public final ConfigEntry<String> activity;
+    public final ConfigEntry<Long> channelId;
 
-    public final StringConfigEntry token = builder.defineString("token");
-    public final BoolConfigEntry autoStart = builder.defineBool("autoStart", true);
+    public BotConfig(ConfigBuilder builder) {
+        this.token = builder.stringEntry("token", "",
+                "Token of the Discord bot.");
+        this.autoStart = builder.booleanEntry("autoStart", true,
+                "Whether to automatically start the Discord bot on server start.");
+        this.activityType = builder.enumEntry("activityType", ActivityType.PLAYING,
+                "Type of activity the Discord bot should display. Allowed values: NONE, LISTENING, PLAYING, COMPETING, WATCHING, CUSTOM");
+        this.activity = builder.stringEntry("activity", "Minecraft",
+                "The activity the Discord bot should display.");
+        this.channelId = builder.longEntry("channelId", 0L,
+                "ID of the Discord channel the Discord bot should send its messages in.");
+    }
 
-    public ConfigDefinition build() {
-        return builder.build();
+    public enum ActivityType {
+        NONE((s) -> null),
+        LISTENING(Activity::listening),
+        PLAYING(Activity::playing),
+        COMPETING(Activity::competing),
+        WATCHING(Activity::watching),
+        CUSTOM(Activity::customStatus);
+
+        private final Function<String, @Nullable Activity> activityFunction;
+
+        ActivityType(Function<String, @Nullable Activity> activityFunction) {
+            this.activityFunction = activityFunction;
+        }
+
+        @Nullable
+        public Activity createActivity(@NonNull String activityName) {
+            return this.activityFunction.apply(activityName);
+        }
     }
 }