+++ src/main/java/com/github/command17/fabricdiscordbridge/config/ModConfig.java
@@ -1,78 +1,53 @@
 package com.github.command17.fabricdiscordbridge.config;
 
-import com.github.command17.fabricdiscordbridge.config.entry.*;
-import net.dv8tion.jda.api.entities.Activity;
-import net.minecraft.util.StringRepresentable;
-import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
-
-import java.util.function.Function;
+import de.maxhenkel.configbuilder.ConfigBuilder;
+import de.maxhenkel.configbuilder.entry.ConfigEntry;
 
 public class ModConfig {
-    private final ConfigBuilder builder = new ConfigBuilder("fabricdiscordbridge");
+    public final ConfigEntry<Boolean> enabled;
+    public final ConfigEntry<Boolean> serverStartMsgEnabled;
+    public final ConfigEntry<Boolean> serverStopMsgEnabled;
+    public final ConfigEntry<Boolean> playerJoinMsgEnabled;
+    public final ConfigEntry<Boolean> playerLeaveMsgEnabled;
+    public final ConfigEntry<Boolean> playerAwardAdvancementMsgEnabled;
+    public final ConfigEntry<Boolean> playerDiedMsgEnabled;
 
-    public final BoolConfigEntry enabled = builder.defineBool("enabled", true);
+    public final ConfigEntry<Boolean> m2dEnabled;
+    public final ConfigEntry<Boolean> m2dSendCommandOnly;
+    public final ConfigEntry<String> m2dCommandMsg;
 
-    public final EnumConfigEntry<BotActivityType> activityType = builder.defineEnum("discord.activityType", BotActivityType.NONE);
-    public final StringConfigEntry activityName = builder.defineString("discord.activityName");
+    public final ConfigEntry<Boolean> d2mEnabled;
+    public final ConfigEntry<Integer> d2mMaxMsgLength;
+    public final ConfigEntry<String> d2mMsg;
 
-    public final LongConfigEntry channelId = builder.defineLong("discord.channelId", 0);
+    public ModConfig(ConfigBuilder builder) {
+        this.enabled = builder.booleanEntry("enabled", true,
+                "Enables the Discord bot.");
+        this.serverStartMsgEnabled = builder.booleanEntry("discord.serverStartMsgEnabled", true,
+                "Enables the server start message.");
+        this.serverStopMsgEnabled = builder.booleanEntry("discord.serverStopMsgEnabled", true,
+                "Enables the server stop message.");
+        this.playerJoinMsgEnabled = builder.booleanEntry("discord.playerJoinMsgEnabled", true,
+                "Enables the player join message.");
+        this.playerLeaveMsgEnabled = builder.booleanEntry("discord.playerLeaveMsgEnabled", true,
+                "Enables the player leave message.");
+        this.playerAwardAdvancementMsgEnabled = builder.booleanEntry("discord.playerAwardAdvancementMsgEnabled", true,
+                "Enables the player award advancement message.");
+        this.playerDiedMsgEnabled = builder.booleanEntry("discord.playerDiedMsgEnabled", true,
+                "Enables the player died message.");
 
-    public final BoolConfigEntry discordPlayerJoinMsgEnabled = builder.defineBool("discord.playerJoinMsg.enabled", true);
-    public final IntConfigEntry discordPlayerJoinMsgColor = builder.defineInt("discord.playerJoinMsg.colorInt", 16777045);
-    public final StringConfigEntry discordPlayerJoinMsg = builder.defineString("discord.playerJoinMsg", "**%player% has joined the server!**");
+        this.m2dEnabled = builder.booleanEntry("m2d.enabled", true,
+                "Enables the Minecraft to Discord chat bridge.");
+        this.m2dSendCommandOnly = builder.booleanEntry("m2d.sendCommandOnly", false,
+                "If true, messages from chat will only be sent to Discord via a command.");
+        this.m2dCommandMsg = builder.stringEntry("m2d.commandMsg", "§7[{player} -> DISCORD]: {message}",
+                "Message to display in Minecraft when sending a message via command to Discord.\nSupported placeholders: player, playerDisplay, playerUuid, message");
 
-    public final BoolConfigEntry discordPlayerLeaveMsgEnabled = builder.defineBool("discord.playerLeaveMsg.enabled", true);
-    public final IntConfigEntry discordPlayerLeaveMsgColor = builder.defineInt("discord.playerLeaveMsg.colorInt", 16777045);
-    public final StringConfigEntry discordPlayerLeaveMsg = builder.defineString("discord.playerLeaveMsg", "**%player% has left the server!**");
-
-    public final BoolConfigEntry discordPlayerRewardedAdvancementMsgEnabled = builder.defineBool("discord.playerRewardedAdvancementMsg.enabled", true);
-    public final IntConfigEntry discordPlayerRewardedAdvancementMsgColor = builder.defineInt("discord.playerRewardedAdvancementMsg.colorInt", 5635925);
-    public final StringConfigEntry discordPlayerRewardedAdvancementMsg = builder.defineString("discord.playerRewardedAdvancementMsg", "**%player%** completed **%advancementTitle%**!");
-
-    public final BoolConfigEntry discordPlayerDiedMsgEnabled = builder.defineBool("discord.playerDiedMsg.enabled", true);
-    public final IntConfigEntry discordPlayerDiedMsgColor = builder.defineInt("discord.playerDiedMsg.colorInt", 16733525);
-    public final StringConfigEntry discordPlayerDiedMsg = builder.defineString("discord.playerDiedMsg", "**%message%!**");
-
-    public final BoolConfigEntry mcToDcEnabled = builder.defineBool("minecraftToDiscordChat.enabled", true);
-    public final BoolConfigEntry mcToDcSendAsEmbed = builder.defineBool("minecraftToDiscordChat.sendAsEmbed", false);
-    public final BoolConfigEntry mcToDcCommandOnly = builder.defineBool("minecraftToDiscordChat.sendWithCommandOnly", false);
-    public final StringConfigEntry mcToDcCommandMsg = builder.defineString("minecraftToDiscordChat.commandMsg", "<color gray>[%player% -> DISCORD]: %message%</color>");
-    public final StringConfigEntry mcToDcMsg = builder.defineString("minecraftToDiscordChat.msg", "**%player%:** %message%");
-
-    public final BoolConfigEntry dcToMcEnabled = builder.defineBool("discordToMinecraftChat.enabled", true);
-    public final IntConfigEntry dcToMcMaxMsgLength = builder.defineInt("discordToMinecraftChat.maxMsgLength", 256, 0, 4000);
-    public final StringConfigEntry dcToMcMsg = builder.defineString("discordToMinecraftChat.msg", "<color aqua>[DISCORD]</color> <color blue><%displayUser%> %message%</color>");
-
-    public ConfigDefinition build() {
-        return builder.build();
-    }
-
-    public enum BotActivityType implements StringRepresentable {
-        NONE("none", (s) -> null),
-        LISTENING("listening", Activity::listening),
-        PLAYING("playing", Activity::playing),
-        COMPETING("competing", Activity::competing),
-        WATCHING("watching", Activity::watching),
-        CUSTOM("custom", Activity::customStatus);
-
-        private final String name;
-        private final Function<String, @Nullable Activity> activityFunction;
-
-        BotActivityType(@NotNull String name, Function<String, @Nullable Activity> activityFunction) {
-            this.name = name;
-            this.activityFunction = activityFunction;
-        }
-
-        @Nullable
-        public Activity createActivity(@NotNull String activityName) {
-            return this.activityFunction.apply(activityName);
-        }
-
-        @NotNull
-        @Override
-        public String getSerializedName() {
-            return this.name;
-        }
+        this.d2mEnabled = builder.booleanEntry("d2m.enabled", true,
+                "Enables the Discord to Minecraft chat bridge.");
+        this.d2mMaxMsgLength = builder.integerEntry("d2m.maxMsgLength", 256, 0, 4000,
+                "Maximum Discord message length to be displayed in Minecraft.");
+        this.d2mMsg = builder.stringEntry("d2m.msg", "§9[DISCORD] <{discordUserDisplay}>§r {message}",
+                "Message to display in Minecraft when sending a message via Discord.\nSupported placeholders: discordUser, discordUserDisplay, message");
     }
 }