+++ src/main/java/com/github/command17/fabricdiscordbridge/event/ModEvents.java
@@ -1,16 +1,14 @@
 package com.github.command17.fabricdiscordbridge.event;
 
-import com.github.command17.fabricdiscordbridge.DiscordBot;
 import com.github.command17.fabricdiscordbridge.FabricDiscordBridge;
-import com.github.command17.fabricdiscordbridge.util.TextPlaceholderHelper;
-import eu.pb4.placeholders.api.ParserContext;
-import eu.pb4.placeholders.api.parsers.TagLikeParser;
+import com.github.command17.fabricdiscordbridge.config.message.MessageConfigs;
+import com.github.command17.fabricdiscordbridge.util.StringPlaceholder;
+import net.dv8tion.jda.api.utils.MarkdownSanitizer;
 import net.fabricmc.fabric.api.entity.event.v1.ServerLivingEntityEvents;
 import net.fabricmc.fabric.api.entity.event.v1.ServerPlayerEvents;
 import net.fabricmc.fabric.api.event.lifecycle.v1.ServerLifecycleEvents;
 import net.fabricmc.fabric.api.message.v1.ServerMessageEvents;
 import net.minecraft.network.chat.ChatType;
-import net.minecraft.network.chat.Component;
 import net.minecraft.network.chat.PlayerChatMessage;
 import net.minecraft.server.MinecraftServer;
 import net.minecraft.server.level.ServerPlayer;
@@ -18,110 +16,74 @@ import net.minecraft.world.damagesource.DamageSource;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.player.Player;
 
-public final class ModEvents {
-    private static final TextPlaceholderHelper placeholderHelper = new TextPlaceholderHelper(TagLikeParser.PLACEHOLDER);
+import java.util.Map;
 
+public final class ModEvents {
     private static void onServerStarting(MinecraftServer server) {
         if (!FabricDiscordBridge.BOT_CONFIG.autoStart.get()) {
             return;
         }
 
         FabricDiscordBridge.createDiscordBot();
+
+        if (FabricDiscordBridge.CONFIG.serverStartMsgEnabled.get()) {
+            FabricDiscordBridge.withDiscordBot((bot) -> MessageConfigs.SERVER_START.send(bot, null));
+        }
     }
 
     private static void onServerStopped(MinecraftServer server) {
+        if (FabricDiscordBridge.CONFIG.serverStopMsgEnabled.get()) {
+            FabricDiscordBridge.withDiscordBot((bot) -> MessageConfigs.SERVER_STOP.send(bot, null));
+        }
+
         FabricDiscordBridge.shutdownDiscordBot();
     }
 
     private static void onPlayerJoin(ServerPlayer player) {
-        if (!FabricDiscordBridge.CONFIG.discordPlayerJoinMsgEnabled.get()) {
+        if (!FabricDiscordBridge.CONFIG.playerJoinMsgEnabled.get()) {
             return;
         }
 
         FabricDiscordBridge.withDiscordBot((bot) -> {
-            TextPlaceholderHelper.setPlayerPlaceholderValues(placeholderHelper, player);
-            Component msg = placeholderHelper.getParser().parseText(
-                    FabricDiscordBridge.CONFIG.discordPlayerJoinMsg.get(),
-                    ParserContext.of()
-            );
-
-            bot.sendEmbed(
-                    DiscordBot.createSimpleColoredEmbed(
-                            msg.getString(),
-                            FabricDiscordBridge.CONFIG.discordPlayerJoinMsgColor.get()
-                    )
-            );
-
-            placeholderHelper.clearPlaceholderValues();
+            Map<String, String> placeholders = StringPlaceholder.combinePlayerPlaceholders(player);
+            MessageConfigs.PLAYER_JOIN.send(bot, placeholders);
         });
     }
 
     private static void onPlayerLeave(ServerPlayer player) {
-        if (!FabricDiscordBridge.CONFIG.discordPlayerLeaveMsgEnabled.get()) {
+        if (!FabricDiscordBridge.CONFIG.playerLeaveMsgEnabled.get()) {
             return;
         }
 
         FabricDiscordBridge.withDiscordBot((bot) -> {
-            TextPlaceholderHelper.setPlayerPlaceholderValues(placeholderHelper, player);
-            Component msg = placeholderHelper.getParser().parseText(
-                    FabricDiscordBridge.CONFIG.discordPlayerLeaveMsg.get(),
-                    ParserContext.of()
-            );
-
-            bot.sendEmbed(
-                    DiscordBot.createSimpleColoredEmbed(
-                            msg.getString(),
-                            FabricDiscordBridge.CONFIG.discordPlayerLeaveMsgColor.get()
-                    )
-            );
-
-            placeholderHelper.clearPlaceholderValues();
+            Map<String, String> placeholders = StringPlaceholder.combinePlayerPlaceholders(player);
+            MessageConfigs.PLAYER_LEAVE.send(bot, placeholders);
         });
     }
 
     private static void onChat(PlayerChatMessage chatMessage, ServerPlayer player, ChatType.Bound bound) {
-        if (!FabricDiscordBridge.CONFIG.mcToDcEnabled.get() || FabricDiscordBridge.CONFIG.mcToDcCommandOnly.get()) {
+        if (!FabricDiscordBridge.CONFIG.m2dEnabled.get() || FabricDiscordBridge.CONFIG.m2dSendCommandOnly.get()) {
             return;
         }
 
         FabricDiscordBridge.getDiscordBot().ifPresent((bot) -> {
-            placeholderHelper.setPlaceholderValue("message", chatMessage.decoratedContent());
-            TextPlaceholderHelper.setPlayerPlaceholderValues(placeholderHelper, player);
-            Component msg = placeholderHelper.getParser().parseText(
-                    FabricDiscordBridge.CONFIG.mcToDcMsg.get(),
-                    ParserContext.of()
-            );
-
-            if (FabricDiscordBridge.CONFIG.mcToDcSendAsEmbed.get()) {
-                bot.sendEmbed(DiscordBot.createSimpleEmbed(player.getPlainTextName(), msg.getString()));
-            } else {
-                bot.sendComponent(msg);
-            }
-
-            placeholderHelper.clearPlaceholderValues();
+            String sanitizedChatMessage = MarkdownSanitizer.sanitize(chatMessage.decoratedContent().getString());
+            StringPlaceholder messagePlaceholder = StringPlaceholder.message(sanitizedChatMessage);
+            Map<String, String> placeholders = StringPlaceholder.combineWithPlayerPlaceholders(player, messagePlaceholder);
+            MessageConfigs.PLAYER_CHAT.send(bot, placeholders);
         });
     }
 
     private static void onLivingEntityDied(LivingEntity entity, DamageSource damageSource) {
-        if (!FabricDiscordBridge.CONFIG.discordPlayerDiedMsgEnabled.get()) {
+        if (!FabricDiscordBridge.CONFIG.playerDiedMsgEnabled.get()) {
             return;
         }
 
         if (entity instanceof Player player) {
             FabricDiscordBridge.withDiscordBot((bot) -> {
-                placeholderHelper.setPlaceholderValue("message", damageSource.getLocalizedDeathMessage(player));
-                TextPlaceholderHelper.setPlayerPlaceholderValues(placeholderHelper, player);
-                Component msg = placeholderHelper.getParser().parseText(
-                        FabricDiscordBridge.CONFIG.discordPlayerDiedMsg.get(),
-                        ParserContext.of()
-                );
-
-                bot.sendEmbed(DiscordBot.createSimpleColoredEmbed(
-                        msg.getString(),
-                        FabricDiscordBridge.CONFIG.discordPlayerDiedMsgColor.get()
-                ));
-
-                placeholderHelper.clearPlaceholderValues();
+                StringPlaceholder messagePlaceholder = StringPlaceholder.message(damageSource.getLocalizedDeathMessage(player).getString());
+                Map<String, String> placeholders = StringPlaceholder.combineWithPlayerPlaceholders(player, messagePlaceholder);
+                MessageConfigs.PLAYER_DEATH.send(bot, placeholders);
             });
         }
     }