+++ src/main/java/com/github/command17/fabricdiscordbridge/mixin/PlayerAdvancementsMixin.java
@@ -1,29 +1,23 @@
 package com.github.command17.fabricdiscordbridge.mixin;
 
-import com.github.command17.fabricdiscordbridge.DiscordBot;
 import com.github.command17.fabricdiscordbridge.FabricDiscordBridge;
-import com.github.command17.fabricdiscordbridge.util.TextPlaceholderHelper;
-import eu.pb4.placeholders.api.ParserContext;
-import eu.pb4.placeholders.api.PlaceholderContext;
-import eu.pb4.placeholders.api.parsers.TagLikeParser;
+import com.github.command17.fabricdiscordbridge.config.message.MessageConfigs;
+import com.github.command17.fabricdiscordbridge.util.StringPlaceholder;
 import net.minecraft.advancements.AdvancementHolder;
 import net.minecraft.advancements.DisplayInfo;
-import net.minecraft.network.chat.Component;
 import net.minecraft.server.PlayerAdvancements;
 import net.minecraft.server.level.ServerPlayer;
-import net.minecraft.world.level.GameRules;
+import net.minecraft.world.level.gamerules.GameRules;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
-import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
+import java.util.Map;
+
 @Mixin(PlayerAdvancements.class)
 public class PlayerAdvancementsMixin {
-    @Unique
-    private static final TextPlaceholderHelper placeholderHelper = new TextPlaceholderHelper(TagLikeParser.PLACEHOLDER);
-
     @Shadow
     private ServerPlayer player;
 
@@ -35,24 +29,22 @@ public class PlayerAdvancementsMixin {
         }
 
         if (displayInfo.shouldAnnounceChat()
-                && this.player.level().getGameRules().getBoolean(GameRules.RULE_ANNOUNCE_ADVANCEMENTS)
-                && FabricDiscordBridge.CONFIG.discordPlayerRewardedAdvancementMsgEnabled.get()) {
+                && this.player.level().getGameRules().get(GameRules.SHOW_ADVANCEMENT_MESSAGES)
+                && FabricDiscordBridge.CONFIG.playerAwardAdvancementMsgEnabled.get()) {
             FabricDiscordBridge.withDiscordBot((bot) -> {
-                placeholderHelper.setPlaceholderValue("advancementTitle", displayInfo.getTitle());
-                placeholderHelper.setPlaceholderValue("advancementDescription", displayInfo.getDescription());
-                placeholderHelper.setPlaceholderValue("advancementType", displayInfo.getType().getDisplayName());
-                TextPlaceholderHelper.setPlayerPlaceholderValues(placeholderHelper, player);
-                Component msg = placeholderHelper.getParser().parseText(
-                        FabricDiscordBridge.CONFIG.discordPlayerRewardedAdvancementMsg.get(),
-                        ParserContext.of()
+                StringPlaceholder advancementTitlePlaceholder = new StringPlaceholder("advancementTitle", displayInfo.getTitle().getString());
+                StringPlaceholder advancementDescPlaceholder = new StringPlaceholder("advancementDesc", displayInfo.getDescription().getString());
+                StringPlaceholder advancementTypePlaceholder = new StringPlaceholder("advancementType", displayInfo.getType().getSerializedName());
+                StringPlaceholder advancementTypeCapitalPlaceholder = new StringPlaceholder("advancementDisplay", displayInfo.getType().getDisplayName().getString());
+                Map<String, String> placeholders = StringPlaceholder.combineWithPlayerPlaceholders(
+                        player,
+                        advancementTitlePlaceholder,
+                        advancementDescPlaceholder,
+                        advancementTypePlaceholder,
+                        advancementTypeCapitalPlaceholder
                 );
 
-                bot.sendEmbed(DiscordBot.createSimpleColoredEmbed(
-                        msg.getString(),
-                        FabricDiscordBridge.CONFIG.discordPlayerRewardedAdvancementMsgColor.get()
-                ));
-
-                placeholderHelper.clearPlaceholderValues();
+                MessageConfigs.PLAYER_AWARD_ADVANCEMENT.send(bot, placeholders);
             });
         }
     }